Move bugzilla code to a submodule
